Moved the switch responsibilities to the DeviceBar class, not sure about this decission though
